# Filter down objects before doing heavy computations on condition scenarios