*Issue #, if available:*

*Description of changes:*
- Filter down objects before running computationally heavy condition scenarios

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.