@kddejong 2020-01-17
before
`cfn-lint test4.yaml  14.03s user 4.13s system 98% cpu 18.404 total`
after
`cfn-lint test4.yaml  0.96s user 0.20s system 86% cpu 1.335 total`

This isn't a full fix for possible performance issues but is a significant improvement for certain scenarios.